Changed the library Refresh method to not try to process my entire library in parallel. I think it actually runs faster serially without all the locks now.
Fixed a bug where exiting the application while the library is refreshing caused an exception.
Fixed a bug where the basic visualizer was in some cases causing a NullReferenceException.

Changed it so that when the Play button is pressed, it will restart the currently playing song (that's how winamp does it...and I like it that way.)

Tried a ton of crap to fix the AccessViolationExceptions. It's most certainly FFMPEG library code. There are some people complaining about something sort of like it online...the mp3 decoding algorithm they use is buggy, apparently. Recommend we switch to mpg123 or try to find a different version of FFMPEG to use for our decoder.

Also, AccessViolationExceptions can't (apparently) be caught.