Fixed a bug where the generator didn't work. It wasn't properly copying all files needed by each plugin. When testing the application from the release directory after generating, it was reading needed files from the bin directory (I think).
Anyway, projects have been restructured so that each plugin compiles to its own directory. 

The generator scans the directories for plugin DLLs and records their path.

When copying, it copies the directory of the plugin DLL and all subdirectories and files.

Works perfectly for me, tell me if this change conflicts with our documentation and/or causes problems for anybody else.