Fixed a bug where in debug mode, some DLLs would be scanned and throw a ReflectionTypeLoadException.
Fixed a bug in AudioEngineTest where a property formerly known as Plugins was being referred to as such, despite a refactor to the name AllPlugins.

Fixed a bug in Library where iterating over a collection with foreach was causing a SomethingModificationException or something during serialization. Maybe Concurrent. Anyway, it uses for loops now and it's fine.

Changed LibraryConfigurationDialog so that it only refreshes the library if you actually click save. Since refreshing the library can take quite a bit of time, if the user doesn't actually save any changes, it shouldn't be updated!

Changed LibraryCollectionView so that it no longer updates the tree display as it finds artists. Despite being neat at first, it slows things down library refreshing and deserialization considerably.

Tweaked the visualizer.